migrate dependency sorting helper functions to dependency-walker for context
